+++ CHANGELOG.md
@@ -1,3 +1,62 @@
+## 0.79.0 (2026-02-26)
+
+### 🚀 Features
+
+- #480 front-end route and component structure for AgentOS ([#524](https://github.com/whoz-oss/coday/pull/524), [#480](https://github.com/whoz-oss/coday/issues/480))
+
+### ❤️ Thank You
+
+- vincent-audibert-whoz
+
+## 0.78.0 (2026-02-25)
+
+### 🚀 Features
+
+- #444 handle integrations by name and type ([#517](https://github.com/whoz-oss/coday/pull/517), [#444](https://github.com/whoz-oss/coday/issues/444))
+
+### ❤️ Thank You
+
+- vincent-audibert-whoz
+
+## 0.77.0 (2026-02-25)
+
+### 🚀 Features
+
+- #443 agent editor ([#513](https://github.com/whoz-oss/coday/pull/513), [#443](https://github.com/whoz-oss/coday/issues/443))
+
+### ❤️ Thank You
+
+- vincent-audibert-whoz
+
+## 0.76.0 (2026-02-25)
+
+### 🚀 Features
+
+- #520 propagate stop signal and handle interrupted delegations ([#521](https://github.com/whoz-oss/coday/pull/521), [#520](https://github.com/whoz-oss/coday/issues/520))
+
+### ❤️ Thank You
+
+- vincent-audibert-whoz
+
+## 0.75.0 (2026-02-25)
+
+### 🚀 Features
+
+- wz-28725: KLogging for filesystem plugin ([b736345](https://github.com/whoz-oss/coday/commit/b736345))
+- wz-28725: KLogging for datetime plugin ([7c68553](https://github.com/whoz-oss/coday/commit/7c68553))
+- wz-28725: postman collection for cases endpoints ([bab724a](https://github.com/whoz-oss/coday/commit/bab724a))
+- wz-28725: build plugins before running agentOS ([940d1c6](https://github.com/whoz-oss/coday/commit/940d1c6))
+- wz-28725: build plugins before running agentOS ([#507](https://github.com/whoz-oss/coday/pull/507))
+- wz-28725: KLogging for plugins ([#506](https://github.com/whoz-oss/coday/pull/506))
+- #482 add Frontay frontend agent, split Runnay into Nxay/Gradlay ([#523](https://github.com/whoz-oss/coday/pull/523), [#482](https://github.com/whoz-oss/coday/issues/482))
+- #391 better file tools ([#522](https://github.com/whoz-oss/coday/pull/522), [#391](https://github.com/whoz-oss/coday/issues/391))
+
+### ❤️ Thank You
+
+- Thomas MARTIN @thomasmartin-whoz
+- thomas.martin @thomasmartin-whoz
+- vincent-audibert-whoz
+
 ## 0.74.0 (2026-02-19)
 
 ### 🚀 Features